Refactor Response and MultiResponse
Response: Add an instance variable that controls whether the response's
keys should be symbolized. This will be used in the upcoming api_version
key, that enables a backwards compatible format.

MultiResponse: Having an attr_reader for `body` and a manually defined
method with the same name, that is actually no reader for `@body`, is a
bit confusing. Therefore, I renamed the method and made use of the
attr_reader more "explicit".