Add optional api_version argument
This allows to choose an older response format for classifications and
extractions and should help migrating to the new client.

Also, this will allow to add a `train` method vor v2 models, which had
been removed with the new gem. This method will just use the v2 API
under the hood.