Add train endpoint for classifiers
This endpoint uses the v2 API endpoint under the hood. This too will
help with migrating to the new gem, as this functionality is necessary
to manage v2 classifiers using the v3 gem.

Should this endpoint be triggered on a v3 model, it's the back-end's
task to respond with an error.